requirements.txt
torch
numpy
opencv-python
scipy
facenet-pytorch
Pillow
fastapi
python-multipart
uvicorn